quantum-lens-backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
mysql-connector-python==8.0.26
pandas==1.3.3
python-dotenv==0.19.0
pydantic==1.8.2
requests==2.26.0
aiohttp==3.8.1
cryptography==3.4.8
python-jose[cryptography]==3.3.0 